Changed to doubles instead of ints Added CPI test Removed ASGS from datasets file Removed state from queryInputs when ASGS was being used

+++ src/main/java/naturallanguagedotstat/Service.java
@@ -128,7 +128,7 @@ public class Service {
 			urlToRead = queryBuilder.build();
 
 			String data = null;
-			int result = -1;
+			double result = -1;
 
 			if(!LocalTest.test){
 				data = Utils.httpGET(urlToRead);
